python_backend/pyproject.toml
[tool.poetry]
name = "sncf_max_tickets_checker"
version = "0.1.0"
description = "A project to check the maximum available SNCF tickets using HTTP requests"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
aiohttp = "^3.8.5"
asyncio = "^3.4.3"
httpx = "^0.24.1"
python-dotenv = "^0.19.2"
fastapi = "^0.70.0"
uvicorn = "^0.15.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
black = "^23.3.0"
flake8 = "^6.1.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
